# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SentinAI multi-model AI API gateway
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ModelKeys {
  openai?: string;
  gemini?: string;
  anthropic?: string;
}

export type PiiMatchType = (typeof PiiMatchType)[keyof typeof PiiMatchType];

export const PiiMatchType = {
  email: "email",
  phone: "phone",
  ssn: "ssn",
} as const;

export interface PiiMatch {
  type: PiiMatchType;
  value: string;
}

export interface PiiBlockedResponse {
  error: string;
  piiMatches: PiiMatch[];
}

export interface CompareRequest {
  prompt: string;
  tenantId?: string;
  modelKeys?: ModelKeys;
}

export type CompareModelResultStatus =
  (typeof CompareModelResultStatus)[keyof typeof CompareModelResultStatus];

export const CompareModelResultStatus = {
  success: "success",
  error: "error",
} as const;

export interface CompareModelResult {
  /** @nullable */
  response?: string | null;
  tokens: number;
  cost: number;
  latencyMs: number;
  status: CompareModelResultStatus;
  /** @nullable */
  error?: string | null;
}

export type CompareResponseResults = {
  openai: CompareModelResult;
  gemini: CompareModelResult;
  claude: CompareModelResult;
  "claude-opus": CompareModelResult;
};

export interface CompareResponse {
  results: CompareResponseResults;
}

export type ChatRequestModel =
  (typeof ChatRequestModel)[keyof typeof ChatRequestModel];

export const ChatRequestModel = {
  openai: "openai",
  gemini: "gemini",
  claude: "claude",
  "claude-opus": "claude-opus",
} as const;

export interface ChatRequest {
  prompt: string;
  model: ChatRequestModel;
  /** @nullable */
  apiKey?: string | null;
  tenantId?: string;
  modelKeys?: ModelKeys;
}

export type ChatResponseStatus =
  (typeof ChatResponseStatus)[keyof typeof ChatResponseStatus];

export const ChatResponseStatus = {
  success: "success",
  fallback: "fallback",
  blocked: "blocked",
  error: "error",
} as const;

export interface ChatResponse {
  response: string;
  model: string;
  modelUsed: string;
  fallback: boolean;
  tokens: number;
  cost: number;
  latencyMs: number;
  status: ChatResponseStatus;
  /** @nullable */
  errorMessage?: string | null;
}

export interface ModelStats {
  requests: number;
  tokens: number;
  cost: number;
  totalLatencyMs: number;
  avgLatencyMs: number;
}

export type StatsResponseModels = {
  openai: ModelStats;
  gemini: ModelStats;
  claude: ModelStats;
  "claude-opus": ModelStats;
};

export interface StatsResponse {
  totalRequests: number;
  blockedRequests: number;
  models: StatsResponseModels;
}

export type LogEntryStatus =
  (typeof LogEntryStatus)[keyof typeof LogEntryStatus];

export const LogEntryStatus = {
  success: "success",
  fallback: "fallback",
  blocked: "blocked",
  error: "error",
} as const;

export interface LogEntry {
  id: string;
  timestamp: string;
  model: string;
  modelUsed: string;
  tokens: number;
  cost: number;
  latencyMs: number;
  status: LogEntryStatus;
  promptSnippet: string;
  /** @nullable */
  responseText?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  tenantId: string;
}

export interface LogsResponse {
  logs: LogEntry[];
}

export interface TenantSummary {
  tenantId: string;
  totalRequests: number;
  blockedRequests: number;
  totalCost: number;
  totalTokens: number;
}

export interface TenantsResponse {
  tenants: TenantSummary[];
}

export interface ChatFailedResponse {
  error: string;
  /** @nullable */
  suggestedFallback?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type GetStatsParams = {
  /**
   * Tenant ID to filter stats for. Returns aggregate across all tenants if omitted.
   */
  tenantId?: string;
};

export type GetLogsParams = {
  /**
   * Tenant ID to filter logs for. Returns all logs if omitted.
   */
  tenantId?: string;
};