# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SentinAI multi-model AI API gateway
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /compare:
    post:
      operationId: compare
      tags: [chat]
      summary: Send the same prompt to all models simultaneously
      description: Runs the prompt against all four models in parallel and returns results side-by-side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompareRequest"
      responses:
        "200":
          description: Results from all models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompareResponse"
        "400":
          description: PII detected in prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PiiBlockedResponse"
...
components:
  schemas:
...
    ModelKeys:
      type: object
      properties:
        openai:
          type: string
        gemini:
          type: string
        anthropic:
          type: string
    PiiMatch:
      type: object
      properties:
        type:
          type: string
          enum: [email, phone, ssn]
        value:
          type: string
      required:
        - type
        - value
    PiiBlockedResponse:
      type: object
      properties:
        error:
          type: string
        piiMatches:
          type: array
          items:
            $ref: "#/components/schemas/PiiMatch"
      required:
        - error
        - piiMatches
    CompareRequest:
      type: object
      properties:
        prompt:
          type: string
        tenantId:
          type: string
        modelKeys:
          $ref: "#/components/schemas/ModelKeys"
      required:
        - prompt
    CompareModelResult:
      type: object
      properties:
        response:
          type: ["string", "null"]
        tokens:
          type: integer
        cost:
          type: number
        latencyMs:
          type: integer
        status:
          type: string
          enum: [success, error]
        error:
          type: ["string", "null"]
      required:
        - tokens
        - cost
        - latencyMs
        - status
    CompareResponse:
      type: object
      properties:
        results:
          type: object
          properties:
            openai:
              $ref: "#/components/schemas/CompareModelResult"
            gemini:
              $ref: "#/components/schemas/CompareModelResult"
            claude:
              $ref: "#/components/schemas/CompareModelResult"
            claude-opus:
              $ref: "#/components/schemas/CompareModelResult"
          required:
            - openai
            - gemini
            - claude
            - claude-opus
      required:
        - results
...