# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SentinAI multi-model AI API gateway
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Routes prompt to the selected LLM provider with automatic fallback on failure or rate limit
 * @summary Send a chat prompt to a model
 */
export const ChatBody = zod.object({
  prompt: zod.string(),
  model: zod.enum(["openai", "gemini", "claude", "claude-opus"]),
  apiKey: zod.string().nullish(),
  tenantId: zod.string().optional(),
  modelKeys: zod
    .object({
      openai: zod.string().optional(),
      gemini: zod.string().optional(),
      anthropic: zod.string().optional(),
    })
    .optional(),
});

export const ChatResponse = zod.object({
  response: zod.string(),
  model: zod.string(),
  modelUsed: zod.string(),
  fallback: zod.boolean(),
  tokens: zod.number(),
  cost: zod.number(),
  latencyMs: zod.number(),
  status: zod.enum(["success", "fallback", "blocked", "error"]),
  errorMessage: zod.string().nullish(),
});

/**
 * Runs the prompt against all four models in parallel and returns results side-by-side
 * @summary Send the same prompt to all models simultaneously
 */
export const CompareBody = zod.object({
  prompt: zod.string(),
  tenantId: zod.string().optional(),
  modelKeys: zod
    .object({
      openai: zod.string().optional(),
      gemini: zod.string().optional(),
      anthropic: zod.string().optional(),
    })
    .optional(),
});

export const CompareResponse = zod.object({
  results: zod.object({
    openai: zod.object({
      response: zod.string().nullish(),
      tokens: zod.number(),
      cost: zod.number(),
      latencyMs: zod.number(),
      status: zod.enum(["success", "error"]),
      error: zod.string().nullish(),
    }),
    gemini: zod.object({
      response: zod.string().nullish(),
      tokens: zod.number(),
      cost: zod.number(),
      latencyMs: zod.number(),
      status: zod.enum(["success", "error"]),
      error: zod.string().nullish(),
    }),
    claude: zod.object({
      response: zod.string().nullish(),
      tokens: zod.number(),
      cost: zod.number(),
      latencyMs: zod.number(),
      status: zod.enum(["success", "error"]),
      error: zod.string().nullish(),
    }),
    "claude-opus": zod.object({
      response: zod.string().nullish(),
      tokens: zod.number(),
      cost: zod.number(),
      latencyMs: zod.number(),
      status: zod.enum(["success", "error"]),
      error: zod.string().nullish(),
    }),
  }),
});

/**
 * Returns total requests, per-model stats, costs, and blocked request count. Filtered by tenantId if provided.
 * @summary Get gateway statistics
 */
export const GetStatsQueryParams = zod.object({
  tenantId: zod.coerce
    .string()
    .optional()
    .describe(
      "Tenant ID to filter stats for. Returns aggregate across all tenants if omitted.",
    ),
});

export const GetStatsResponse = zod.object({
  totalRequests: zod.number(),
  blockedRequests: zod.number(),
  models: zod.object({
    openai: zod.object({
      requests: zod.number(),
      tokens: zod.number(),
      cost: zod.number(),
      totalLatencyMs: zod.number(),
      avgLatencyMs: zod.number(),
    }),
    gemini: zod.object({
      requests: zod.number(),
      tokens: zod.number(),
      cost: zod.number(),
      totalLatencyMs: zod.number(),
      avgLatencyMs: zod.number(),
    }),
    claude: zod.object({
      requests: zod.number(),
      tokens: zod.number(),
      cost: zod.number(),
      totalLatencyMs: zod.number(),
      avgLatencyMs: zod.number(),
    }),
    "claude-opus": zod.object({
      requests: zod.number(),
      tokens: zod.number(),
      cost: zod.number(),
      totalLatencyMs: zod.number(),
      avgLatencyMs: zod.number(),
    }),
  }),
});

/**
 * Returns the last 20 request log entries, filtered by tenantId if provided.
 * @summary Get recent traffic logs
 */
export const GetLogsQueryParams = zod.object({
  tenantId: zod.coerce
    .string()
    .optional()
    .describe("Tenant ID to filter logs for. Returns all logs if omitted."),
});

export const GetLogsResponse = zod.object({
  logs: zod.array(
    zod.object({
      id: zod.string(),
      timestamp: zod.string(),
      model: zod.string(),
      modelUsed: zod.string(),
      tokens: zod.number(),
      cost: zod.number(),
      latencyMs: zod.number(),
      status: zod.enum(["success", "fallback", "blocked", "error"]),
      promptSnippet: zod.string(),
      responseText: zod.string().nullish(),
      errorMessage: zod.string().nullish(),
      tenantId: zod.string(),
    }),
  ),
});

/**
 * Returns every known tenant ID with their aggregate request count, cost, and token usage
 * @summary List all tenants with summary stats
 */
export const GetTenantsResponse = zod.object({
  tenants: zod.array(
    zod.object({
      tenantId: zod.string(),
      totalRequests: zod.number(),
      blockedRequests: zod.number(),
      totalCost: zod.number(),
      totalTokens: zod.number(),
    }),
  ),
});