# lib/api-zod/src/generated/types/compareRequest.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SentinAI multi-model AI API gateway
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const CompareRequest = zod.object({
  prompt: zod.string(),
  tenantId: zod.string().optional(),
  modelKeys: zod
    .object({
      openai: zod.string().optional(),
      gemini: zod.string().optional(),
      anthropic: zod.string().optional(),
    })
    .optional(),
});

export type CompareRequest = zod.input<typeof CompareRequest>;